requirements.txt
yapf
albumentations
pretrainedmodels
munch
efficientnet-pytorch==0.6.3
timm==0.3.2
dill==0.3.3
multiprocess==0.70.11.1
pathos==0.2.7
pox==0.2.9
ppft==1.6.6.3
git+https://github.com/qubvel/segmentation_models.pytorch
wandb
git+https://github.com/Project-MONAI/MONAI#egg=monai
lightning_bolts
jsonargparse[signatures]
flatten_dict